EDU-332 refactor: 사후 평가 문제 반환 API 초안 코드 위치 이동
