+++ app/routers/pre_assessment_router.py
@@ -90,6 +90,7 @@ def build_pretest_log(user_id: str, questions: list[dict]):
         "timestamp": datetime.now().isoformat()
     }
 
+
 @router.get("/subject", response_model=List[PreQuestion], response_model_by_alias=False)
 async def get_pretest(user_id: str, subject_id: int):
     user = await get_user(user_id)
@@ -113,74 +114,3 @@ async def get_pretest(user_id: str, subject_id: int):
 
     result = result_generate(selected)
     return result
-
-
-# 사전 평가 문제 반환(임시) -> 사후 평가 문제 반환으로 수정
-@router.get("/subject-tmp", response_model=List[PreQuestion], response_model_by_alias=False)
-async def get_pretest(user_id:str, subject_id: int):
-    # 1. 사용자 정보 조회
-    user = await db.user_profiles.find_one({"user_id": user_id})
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-
-    e_level = user.get("level")
-
-    # 2) techMap 문서 한 번 불러오기
-    mapping = await db.techMap.find_one({})
-    if not mapping:
-        raise HTTPException(status_code=500, detail="Tech map not initialized")
-
-    # 3) subject_name (숫자 키) 조회
-    subject_name = mapping.get(str(subject_id))
-    if not subject_name:
-        raise HTTPException(status_code=404, detail="Subject not found")
-
-    # 4) Korean level (영어 키) 조회
-    difficulties_doc = await db.techMap.find_one({"difficulties": {"$exists": True}})
-    if not difficulties_doc:
-        raise HTTPException(500, "Difficulties mapping not found")
-
-    k_level = difficulties_doc["difficulties"].get(e_level)
-
-    question_count = {}
-    if k_level == "하":
-        question_count["상"], question_count["중"], question_count["하"] = 1, 3, 6
-    elif k_level == "중":
-        question_count["상"], question_count["중"], question_count["하"] = 3, 4, 3
-    elif k_level == "상":
-        question_count["상"], question_count["중"], question_count["하"] = 6, 3, 1
-    else:
-        raise HTTPException(status_code=500, detail="Forbidden attempt occurred")
-
-
-    # 2. 해당 트랙/레벨 문제 로딩
-    all_questions = await db[subject_name].find().to_list(length=1000)
-
-    hard_qs = [q for q in all_questions if q["difficulty"] == "상"]
-    mid_qs = [q for q in all_questions if q["difficulty"] == "중"]
-    easy_qs = [q for q in all_questions if q["difficulty"] == "하"]
-
-    def safe_sample(pool, cnt):
-        if len(pool) < cnt:
-            raise HTTPException(
-                status_code=500,
-                detail=f"Not enough questions: required {cnt}, got {len(pool)}"
-            )
-        return random.sample(pool, cnt)
-
-    selected = []
-    selected += safe_sample(hard_qs, question_count["상"])
-    selected += safe_sample(mid_qs,  question_count["중"])
-    selected += safe_sample(easy_qs, question_count["하"])
-
-    # 4) 최종 섞기
-    random.shuffle(selected)
-
-    # 5) question_id 부여 및 모델 변환
-    results: List[PreQuestion] = []
-    for idx, doc in enumerate(selected, start=1):
-        doc.pop("_id", None)
-        doc["question_id"] = idx
-        results.append(PreQuestion(**doc))
-
-    return results
